When the "Clear" button is pressed, clear the debug view as well as the message view
At the moment, the "Clear" button in the top-right hand of the
screen clears the "messages" view but not the "debug" view.
This fixes that.

`ai` doesn't provide a clear way to clear the `data` variable
which powers the "debug" view, so this takes an alternative
approach: start a new "chat" by incrementing the chat's ID.